@@ -1,6 +1,6 @@
 [metadata]
 name = python-somfy
-version = 1.0.0
+version = 1.1.0
 author = Aleksei Besogonov
 author_email = b@alex.net
 description = Python library for local control of Somfy SDN shades
@@ -25,7 +25,7 @@ packages = find:
 requires =
     pyserial
     pyserial-asyncio
-python_requires = >=3.11
+python_requires = >=3.12
 install_requires =
     typing-extensions;python_version<'3.8'
 
