@@ -4,15 +4,19 @@ import typing
 from asyncio import get_event_loop
 from optparse import OptionParser
 
-from somfy.connector import SomfyConnector, SocketChannel
+from somfy.connector import SomfyConnector, SocketConnectionFactory, fire_and_forget, detect_devices, \
+    try_to_exchange_one
+from somfy.enumutils import hex_enum
 from somfy.messages import SomfyMessage, SomfyMessageId, MASTER_ADDRESS, NodeType, SomfyAddress
 from somfy.payloads import MotorRotationDirectionPayload, PostMotorLimitsPayload, PostMotorPositionPayload, \
-    CtrlMoveToPayload, CtrlMoveToFunction, CtrlStopPayload, NackPayload
-from somfy.serial import SerialChannel
+    CtrlMoveToPayload, CtrlMoveToFunction, CtrlStopPayload, NackPayload, CtrlMoveRelativePayload, RelativeMoveFunction, \
+    PostMotorStatusPayload, SomfyNackReason
+from somfy.serial import SerialConnectionFactory
+from somfy.utils import wait_for_completion, SomfyNackException, move_with_ack
 
 
 async def do_detect(connector: SomfyConnector):
-    res = await connector.detect_devices()
+    res = await detect_devices(connector)
     for d, t in res:
         print("%s\t%s" % (d, t))
 
@@ -22,21 +26,23 @@ async def do_info(connector: SomfyConnector, opts):
         raise Exception("No --addr specified")
     addr = SomfyAddress.make(opts.addr)
 
-    reply = await connector.exchange_one(addr, SomfyMessageId.GET_MOTOR_ROTATION_DIRECTION,
-                                         SomfyMessageId.POST_MOTOR_ROTATION_DIRECTION)
+    reply = await try_to_exchange_one(connector, addr, SomfyMessageId.GET_MOTOR_ROTATION_DIRECTION,
+                                      SomfyMessageId.POST_MOTOR_ROTATION_DIRECTION)
     if not reply:
         raise Exception("Failed to get motor rotation direction")
 
     rot = typing.cast(MotorRotationDirectionPayload, reply.payload)
     print("Rotation direction\t%s" % rot.get_direction())
 
-    reply = await connector.exchange_one(addr, SomfyMessageId.GET_MOTOR_LIMITS, SomfyMessageId.POST_MOTOR_LIMITS)
+    reply = await try_to_exchange_one(connector, addr, SomfyMessageId.GET_MOTOR_LIMITS,
+                                      SomfyMessageId.POST_MOTOR_LIMITS)
     if not reply:
         raise Exception("Failed to get motor limits")
     lim = typing.cast(PostMotorLimitsPayload, reply.payload)
     print("Limit (in pulses)\t%d" % lim.get_limit())
 
-    reply = await connector.exchange_one(addr, SomfyMessageId.GET_MOTOR_POSITION, SomfyMessageId.POST_MOTOR_POSITION)
+    reply = await try_to_exchange_one(connector, addr, SomfyMessageId.GET_MOTOR_POSITION,
+                                      SomfyMessageId.POST_MOTOR_POSITION)
     if not reply:
         raise Exception("Failed to get motor position")
     pos = typing.cast(PostMotorPositionPayload, reply.payload)
@@ -46,55 +52,60 @@ async def do_info(connector: SomfyConnector, opts):
     print("Tilt degrees\t%d" % (pos.get_tilt_degrees() or -1))
     print("Intermediate Position\t%d" % (pos.get_ip() or -1))
 
+    reply = await try_to_exchange_one(connector, addr, SomfyMessageId.GET_MOTOR_STATUS,
+                                      SomfyMessageId.POST_MOTOR_STATUS)
+    if not reply:
+        raise Exception("Failed to get motor position")
+    stat = typing.cast(PostMotorStatusPayload, reply.payload)
+    print("Status\t%s" % hex_enum(stat.get_status()))
+    print("Direction\t%s" % hex_enum(stat.get_direction()))
+    print("Command Source\t%s" % hex_enum(stat.get_command_source()))
+    print("Status Cause\t%s" % hex_enum(stat.get_status_cause()))
+
 
 async def do_move(connector: SomfyConnector, opts):
     if not opts.addr:
         raise Exception("No --addr specified")
-    if not opts.percent:
+    if opts.percent is None:
         raise Exception("No --percent specified")
 
     addr = SomfyAddress.make(opts.addr)
-
-    def filter_type(msg: SomfyMessage):
-        passed = msg.from_addr == addr and msg.msgid in [SomfyMessageId.ACK, SomfyMessageId.NACK]
-        return passed, not passed  # Accepted?, continue?
-
     payload = CtrlMoveToPayload.make(func=CtrlMoveToFunction.POSITION_PERCENT, position=int(opts.percent))
     sent = SomfyMessage(msgid=SomfyMessageId.CTRL_MOVETO, need_ack=True,
                         from_node_type=NodeType.TYPE_ALL, from_addr=MASTER_ADDRESS,
                         to_node_type=NodeType.TYPE_ALL, to_addr=addr, payload=payload)
+    await move_with_ack(addr, connector, sent)
+
+    def print_pos(p):
+        print("Position: %d%% (pulses: %d), IP=%d" % (p.get_position_percent(), p.get_position_pulses(),
+                                                      p.get_ip() or -1))
+
+    await wait_for_completion(addr, connector, print_pos)
+
+
+async def do_move_ip(connector, opts, move_down):
+    if not opts.addr:
+        raise Exception("No --addr specified")
+
+    addr = SomfyAddress.make(opts.addr)
+    payload = CtrlMoveRelativePayload.make(
+        func=RelativeMoveFunction.MOVE_NEXT_IP_DOWN if move_down else RelativeMoveFunction.MOVE_NEXT_IP_UP, parameter=0)
+    sent = SomfyMessage(msgid=SomfyMessageId.CTRL_MOVE_RELATIVE, need_ack=True,
+                        from_node_type=NodeType.TYPE_ALL, from_addr=MASTER_ADDRESS,
+                        to_node_type=NodeType.TYPE_ALL, to_addr=addr, payload=payload)
     try:
-        messages = await connector.exchange(sent, filter_type)
-        if len(messages) == 0:
-            raise Exception("No ACK or NACK messages")
-    except asyncio.TimeoutError:
-        raise Exception("Move command timed out")
+        await move_with_ack(addr, connector, sent)
+    except SomfyNackException as e:
+        if e.nack().get_nack_code() == SomfyNackReason.NACK_LAST_IP_REACHED:
+            opts.percent = 100 if move_down else 0.0
+            await do_move(connector, opts)
+            return
 
-    if messages[0].msgid == SomfyMessageId.NACK:
-        raise Exception("NACK received, reason: %s" % typing.cast(NackPayload, messages[0].payload).get_nack_code())
-    elif messages[0].msgid != SomfyMessageId.ACK:
-        raise Exception("Move command failed")
+    def print_pos(p):
+        print("Position: %d%% (pulses: %d), IP=%d" % (p.get_position_percent(), p.get_position_pulses(),
+                                                      p.get_ip() or -1))
 
-    await wait_for_completion(addr, connector)
-
-
-async def wait_for_completion(addr, connector):
-    loop = get_event_loop()
-    last_change_time = loop.time()
-    last_pulses = 0
-    # Keep polling while the shades are moving
-    while loop.time() - last_change_time <= 3:
-        reply = await connector.exchange_one(addr, SomfyMessageId.GET_MOTOR_POSITION,
-                                             SomfyMessageId.POST_MOTOR_POSITION)
-        if reply:
-            pos = typing.cast(PostMotorPositionPayload, reply.payload)
-            if pos.get_position_pulses() != last_pulses:
-                last_pulses = pos.get_position_pulses()
-                last_change_time = loop.time()
-            print("Position: %d%% (pulses: %d), IP=%d" % (pos.get_position_percent(), pos.get_position_pulses(),
-                                                          pos.get_ip() or -1))
-
-        await asyncio.sleep(0.5)
+    await wait_for_completion(addr, connector, print_pos)
 
 
 async def do_stop(connector: SomfyConnector, opts):
@@ -102,37 +113,40 @@ async def do_stop(connector: SomfyConnector, opts):
         raise Exception("No --addr specified")
 
     addr = SomfyAddress.make(opts.addr)
-    sent = SomfyMessage(msgid=SomfyMessageId.CTRL_STOP, need_ack=True,
-                        from_node_type=NodeType.TYPE_ALL, from_addr=MASTER_ADDRESS,
-                        to_node_type=NodeType.TYPE_ALL, to_addr=addr, payload=CtrlStopPayload.make())
-    await connector.exchange(sent, lambda o: (False, False))
+    stop_msg = SomfyMessage(msgid=SomfyMessageId.CTRL_STOP, need_ack=True,
+                            from_node_type=NodeType.TYPE_ALL, from_addr=MASTER_ADDRESS,
+                            to_node_type=NodeType.TYPE_ALL, to_addr=addr, payload=CtrlStopPayload.make())
+    await fire_and_forget(connector, stop_msg)
 
 
 async def run(opts, cmd):
     if opts.tcp:
         host, port = opts.tcp.split(":")
-        ch = SocketChannel(host=host, port=port)
+        ch = SocketConnectionFactory(host=host, port=port)
     elif opts.serial:
-        ch = SerialChannel(opts.serial)
+        ch = SerialConnectionFactory(opts.serial)
     else:
         raise Exception("Neither --tcp nor --serial options specified")
 
-    async with ch:
-        async with SomfyConnector(ch) as connector:
-            if cmd == "detect":
-                await do_detect(connector)
-            elif cmd == "info":
-                await do_info(connector, opts)
-            elif cmd == "move":
-                await do_move(connector, opts)
-            elif cmd == "stop":
-                await do_stop(connector, opts)
-            else:
-                raise Exception("Unknown command")
+    async with SomfyConnector(ch) as connector:
+        if cmd == "detect":
+            await do_detect(connector)
+        elif cmd == "info":
+            await do_info(connector, opts)
+        elif cmd == "move":
+            await do_move(connector, opts)
+        elif cmd == "stop":
+            await do_stop(connector, opts)
+        elif cmd == "down_step":
+            await do_move_ip(connector, opts, True)
+        elif cmd == "up_step":
+            await do_move_ip(connector, opts, False)
+        else:
+            raise Exception("Unknown command")
 
 
 if __name__ == '__main__':
-    parser = OptionParser("sdntool.py [options] detect|info|move|stop")
+    parser = OptionParser("sdntool.py [options] detect|info|move|stop|down_step|up_step")
     parser.add_option("--tcp", dest="tcp",
                       help="use the TCP endpoint for the Somfy connection (host:port)")
     parser.add_option("--serial", dest="serial",
