@@ -1,21 +1,17 @@
-import enum
 from enum import IntEnum
-from typing import Any
+from typing import Type
 
 
-class FlexibleEnumMeta(enum.EnumMeta):
-    def __call__(cls, value: Any, *args, **kwargs):
-        try:
-            # attempt to get an enum member
-            return super().__call__(value, *args, **kwargs)
-        except ValueError:
-            # no such member exists, but we don't care
-            return value
-
-
-class EnumWithMissing(IntEnum, metaclass=FlexibleEnumMeta):
+class IntEnumWithStr(IntEnum):
     def __str__(self):
         return self.name
+
+
+def enum_or_int[T](enum_type: Type[T], val: int) -> T | int:
+    try:
+        return enum_type(val)
+    except ValueError:
+        return val
 
 
 def hex_enum(val: IntEnum | int) -> str:
